+++ src/Me/Navigation/Navbar.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { useRouter } from 'next/router';
 import classNames from 'classnames';
 import styled from 'styled-components/macro';
@@ -12,15 +12,19 @@ import IconMentors from '../../assets/me/mentors.svg';
 import IconLogout from '../../assets/me/icon-door-exit.svg';
 import { useUser } from '../../context/userContext/UserContext';
 import { useRoutes } from '../../hooks/useRoutes';
-
+import Tawk from '../Tawk';
 const MenuItem = ({
   icon: Icon,
   label,
   to,
+  setIsTawkShowing,
+  tawkShowing,
 }: {
   icon: string;
   label: string;
   to?: string;
+  setIsTawkShowing?: (arg0: boolean) => void;
+  tawkShowing?: boolean;
 }) => {
   const router = useRouter();
   if (to) {
@@ -36,7 +40,7 @@ const MenuItem = ({
     );
   } else
     return (
-      <button onClick={() => console.log('open the tawk')}>
+      <button onClick={() => setIsTawkShowing(!tawkShowing)}>
         <NavItemDecoration
           className={classNames({ active: router.pathname === to })}
         >
@@ -48,10 +52,13 @@ const MenuItem = ({
 };
 
 const Navbar = () => {
+  const [tawkShowing, setIsTawkShowing] = useState(false);
   const { isAdmin, logout } = useUser();
   const urls = useRoutes();
+  console.log({ tawkShowing });
   return (
     <>
+      {tawkShowing && <Tawk />}
       <Menu>
         <Logo
           src={`${process.env.NEXT_PUBLIC_PUBLIC_URL}/codingcoach-logo-192.png`}
@@ -64,7 +71,12 @@ const Navbar = () => {
           label="Mentorships"
         />
         <MenuItem to={urls.root.get()} icon={IconMentors} label="Mentors" />
-        <MenuItem icon={IconMentors} label="Tawk" />
+        <MenuItem
+          icon={IconMentors}
+          setIsTawkShowing={setIsTawkShowing}
+          tawkShowing={tawkShowing}
+          label="Tak"
+        />
 
         {isAdmin && (
           <MenuItem to={urls.me.admin.get()} icon={IconMentors} label="Admin" />
@@ -144,3 +156,6 @@ const Logout = styled(NavItemDecoration)`
     margin-bottom: 10px;
   }
 `;
+function setIsTawkShowing(arg0: boolean): void {
+  throw new Error('Function not implemented.');
+}
